+++ test/unit/mysql_perf/unit_test_run.sh
@@ -7,6 +7,7 @@
 echo ""
 echo "Unit testing..."
 /usr/bin/python test/unit/mysql_perf/convert_dict.py
+/usr/bin/python test/unit/mysql_perf/create_data_config.py
 /usr/bin/python test/unit/mysql_perf/help_message.py
 /usr/bin/python test/unit/mysql_perf/main.py
 /usr/bin/python test/unit/mysql_perf/mysql_stat.py