+++ test/unit/sonarqube_code_coverage.sh
@@ -9,6 +9,7 @@ coverage erase
 echo ""
 echo "Running unit test modules in conjunction with coverage"
 coverage run -a --source=mysql_perf test/unit/mysql_perf/convert_dict.py
+coverage run -a --source=mysql_perf test/unit/mysql_perf/create_data_config.py
 coverage run -a --source=mysql_perf test/unit/mysql_perf/help_message.py
 coverage run -a --source=mysql_perf test/unit/mysql_perf/main.py
 coverage run -a --source=mysql_perf test/unit/mysql_perf/mysql_stat.py